nosqli/requirements.txt
flask
pymongo
dnspython
pyjwt